requirements.txt
Django==1.7.7
django-appconf==1.0.1
django-notification==1.3.3
django-rosetta==0.7.5
django-user-accounts==1.0.1
ipdb==0.8
ipython==3.0.0
microsofttranslator==0.5
Pillow==2.7.0
polib==1.0.4
pytz==2015.2
recaptcha-client==1.0.6
requests==2.6.0
six==1.9.0
